requirements.txt
adal==1.2.7
applicationinsights==0.11.10
azure-common==1.1.27
azure-core==1.20.1
azure-graphrbac==0.61.1
azure-identity==1.7.0
azure-mgmt-authorization==0.61.0
azure-mgmt-containerregistry==8.2.0
azure-mgmt-core==1.3.0
azure-mgmt-keyvault==9.3.0
azure-mgmt-resource==13.0.0
azure-mgmt-storage==11.2.0
azureml-automl-core==1.36.0
azureml-core==1.36.0.post2
azureml-dataprep==2.24.3
azureml-dataprep-native==38.0.0
azureml-dataprep-rslex==2.0.3
azureml-dataset-runtime==1.36.0
azureml-pipeline==1.36.0
azureml-pipeline-core==1.36.0
azureml-pipeline-steps==1.36.0
azureml-telemetry==1.36.0
azureml-train-automl-client==1.36.0
azureml-train-core==1.36.0
azureml-train-restclients-hyperdrive==1.36.0
backports.tempfile==1.0
backports.weakref==1.0.post1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cloudpickle==1.6.0
contextlib2==21.6.0
cryptography==3.4.8
distro==1.6.0
docker==5.0.3
dotnetcore2==2.1.21
idna==3.3
isodate==0.6.0
jeepney==0.7.1
jmespath==0.10.0
jsonpickle==2.0.0
msal==1.16.0
msal-extensions==0.3.0
msrest==0.6.21
msrestazure==0.6.4
ndg-httpsclient==0.5.1
numpy==1.21.4
oauthlib==3.1.1
pathspec==0.9.0
portalocker==1.7.1
pyarrow==3.0.0
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==20.0.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
SecretStorage==3.3.1
six==1.16.0
urllib3==1.26.7
websocket-client==1.2.1
